Rename Session2_RoomClearing.py to Session3_RoomClearing.py
This Python pseudocode (simplified for demonstrative purposes) presents an example program for a hard-coded room-clearing robot. The exercise tasked participants - with and without familiarity with Python or other coding languages - with identifying 1) the general steps in the robot's algorithm, 2) components that contain ambiguity into which they should inquire further, and in conjunction with the previous, 3) what elements might pose risks of error, unintentional behavior, or outright failure. In pursuit of this, following the first step, participants can dive into the accompanying set of (also pseudocode) libraries.

The exercise provides participants with a grounded example of what issues and risks can face hard-coded systems and their designers, for sake of anchoring later comparison with the risks and challenges of various machine learning approaches. It also contributes in a small way to understanding the challenges facing software engineers, for sake of future communication.